Move type defs to definition module
